requirements.txt
torchaudio
git+https://github.com/m-bain/whisperx
pyannote-audio
pyannote.audio
django
langchain 
langchain-community
langchain-groq
python-dotenv